added copyrigth, documentation to missing methods, and added new method for listing objects
