+++ aws-common/src/commonMain/kotlin/com/estivensh4/aws_kmp/AwsException.kt
@@ -1,3 +1,7 @@
+/*
+ * Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+ */
+
 package com.estivensh4.aws_kmp
 
 class AwsException : Exception {
