+++ aws-s3/src/androidMain/kotlin/com/estivensh4/aws_s3/util/HttpMethodExt.kt
@@ -1,3 +1,7 @@
+/*
+ * Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+ */
+
 package com.estivensh4.aws_s3.util
 
 import com.estivensh4.aws_s3.HttpMethod
