+++ aws-s3/src/androidUnitTest/kotlin/com/estivensh4/aws_s3/AwsS3Test.kt
@@ -1,3 +1,7 @@
+/*
+ * Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+ */
+
 package com.estivensh4.aws_s3
 
 import io.kotest.core.spec.style.FunSpec
