+++ aws-s3/src/commonMain/kotlin/com/estivensh4/aws_s3/AWSS3.kt
@@ -1,14 +1,16 @@
+/*
+ * Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+ */
+
 package com.estivensh4.aws_s3
 
 import com.estivensh4.aws_kmp.AwsException
 
-expect class AwsS3 private constructor(
+expect class AWSS3 private constructor(
     accessKey: String,
     secretKey: String,
     endpoint: String
 ) {
-
-    val endpointAWS: String
 
     /**
      *
@@ -50,14 +52,14 @@ expect class AwsS3 private constructor(
      * @param bucketName The name of the bucket containing the desired object.
      * @param key The key in the specified bucket under which the desired object
      * is stored.
-     * @param expiration The time at which the returned pre-signed URL will
+     * @param expirationInSeconds The time at which the returned pre-signed URL will
      * expire.
      * @return A pre-signed URL which expires at the specified time, and can be
      * used to allow anyone to download the specified object from S3,
      * without exposing the owner's AWS secret access key.
-     * @throws AmazonClientException If there were any problems pre-signing the
+     * @throws AwsException If there were any problems pre-signing the
      * request for the specified S3 object.
-     * @see AwsS3.generatePresignedUrl
+     * @see AWSS3.generatePresignedUrl
      */
     fun generatePresignedUrl(bucketName: String, key: String, expirationInSeconds: Long): String?
 
@@ -101,7 +103,7 @@ expect class AwsS3 private constructor(
      * @param bucketName The name of the bucket containing the desired object.
      * @param key The key in the specified bucket under which the desired object
      * is stored.
-     * @param expiration The time at which the returned pre-signed URL will
+     * @param expirationInSeconds The time at which the returned pre-signed URL will
      * expire.
      * @param method The HTTP method verb to use for this URL
      * @return A pre-signed URL which expires at the specified time, and can be
@@ -109,8 +111,8 @@ expect class AwsS3 private constructor(
      * without exposing the owner's AWS secret access key.
      * @throws AwsException If there were any problems pre-signing the
      * request for the specified S3 object.
-     * @see AwsS3.generatePresignedUrl
-     * @see AwsS3.generatePresignedUrl
+     * @see AWSS3.generatePresignedUrl
+     * @see AWSS3.generatePresignedUrl
      */
     fun generatePresignedUrl(
         bucketName: String,
@@ -168,10 +170,10 @@ expect class AwsS3 private constructor(
      * @return A pre-signed URL that can be used to access an Amazon S3 resource
      * without requiring the user of the URL to know the account's AWS
      * security credentials.
-     * @throws AmazonS3Exception If there were any problems pre-signing the
+     * @throws AwsException If there were any problems pre-signing the
      * request for the Amazon S3 resource.
-     * @see AwsS3.generatePresignedUrl
-     * @see AwsS3.generatePresignedUrl
+     * @see AWSS3.generatePresignedUrl
+     * @see AWSS3.generatePresignedUrl
      */
     fun generatePresignedUrl(generatePresignedUrlRequest: GeneratePresignedUrlRequest): String?
 
@@ -179,7 +181,7 @@ expect class AwsS3 private constructor(
      *
      *
      * Creates a new Amazon S3 bucket with the specified name in the default
-     * (US) region, [Region.US_Standard].
+     * (US) region.
      *
      *
      *
@@ -242,20 +244,164 @@ expect class AwsS3 private constructor(
      */
     suspend fun createBucket(bucketName: String): Bucket
 
+    /**
+     *
+     *
+     * Returns a list of all Amazon S3 buckets that the authenticated sender of
+     * the request owns.
+     *
+     *
+     *
+     * Users must authenticate with a valid AWS Access Key ID that is registered
+     * with Amazon S3. Anonymous requests cannot list buckets, and users cannot
+     * list buckets that they did not create.
+     *
+     *
+     * @return A list of all of the Amazon S3 buckets owned by the authenticated
+     * sender of the request.
+     * @throws AWSS3 If any errors are encountered in the client
+     * while making the request or handling the response.
+     * @see AWSS3.listBuckets
+     */
     suspend fun listBuckets(): List<Bucket>
 
+    /**
+     *
+     *
+     * Deletes the specified bucket. All objects (and all object versions, if
+     * versioning was ever enabled) in the bucket must be deleted before the
+     * bucket itself can be deleted.
+     *
+     *
+     *
+     * Only the owner of a bucket can delete it, regardless of the bucket's
+     * access control policy.
+     *
+     *
+     * @param bucketName The name of the bucket to delete.
+     * @throws AwsException If any errors are encountered in the client
+     * while making the request or handling the response.
+     * @see AWSS3.deleteBucket
+     */
     suspend fun deleteBucket(bucketName: String)
+    /**
+     * Deletes multiple objects in a single bucket from S3.
+     *
+     *
+     * In some cases, some objects will be successfully deleted, while some
+     * attempts will cause an error. If any object in the request cannot be
+     * deleted, this method throws a [AwsException] with
+     * details of the error.
+     *
+     * @param bucketName The name of an existing bucket, to which you have permission.
+     * @param keys The request object containing all options for
+     * deleting multiple objects.
+     * @throws AwsException If any errors occurred in Amazon S3 while
+     * processing the request.
+     */
     suspend fun deleteObjects(bucketName: String, vararg keys: String): DeleteObjectResult
+    /**
+     *
+     *
+     * Uploads the specified file to Amazon S3 under the specified bucket and
+     * key name.
+     *
+     *
+     *
+     * Amazon S3 never stores partial objects; if during this call an exception
+     * wasn't thrown, the entire object was stored.
+     *
+     *
+     *
+     * If you are uploading or accessing [AWS KMS](http://aws.amazon.com/kms/)-encrypted objects, you need to
+     * specify the correct region of the bucket on your client and configure AWS
+     * Signature Version 4 for added security. For more information on how to do
+     * this, see
+     * http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#
+     * specify-signature-version
+     *
+     *
+     *
+     * Using the file extension, Amazon S3 attempts to determine the correct
+     * content type and content disposition to use for the object.
+     *
+     *
+     *
+     * If versioning is enabled for the specified bucket, this operation will
+     * this operation will never overwrite an existing object with the same key,
+     * but will keep the existing object as an older version until that version
+     * is explicitly deleted.
+     *
+     *
+     *
+     * If versioning is not enabled, this operation will overwrite an existing
+     * object with the same key; Amazon S3 will store the last write request.
+     * Amazon S3 does not provide object locking. If Amazon S3 receives multiple
+     * write requests for the same object nearly simultaneously, all of the
+     * objects might be stored. However, a single object will be stored with the
+     * final write request.
+     *
+     *
+     *
+     * When specifying a location constraint when creating a bucket, all objects
+     * added to the bucket are stored in the bucket's region. For example, if
+     * specifying a Europe (EU) region constraint for a bucket, all of that
+     * bucket's objects are stored in EU region.
+     *
+     *
+     *
+     * The specified bucket must already exist and the caller must have
+     * permission to the bucket to upload an object.
+     *
+     *
+     * @param bucketName The name of an existing bucket, to which you have permission.
+     * @param key The key under which to store the specified file.
+     * @param imageFile The file containing the data to be uploaded to Amazon S3.
+     * @return A [PutObjectResult] object containing the information
+     * returned by Amazon S3 for the newly created object.
+     * @throws AwsException If any errors are encountered in the client
+     * while making the request or handling the response.
+     * @see AWSS3.putObject
+     */
     suspend fun putObject(
         bucketName: String,
         key: String,
         imageFile: ImageFile
     ): PutObjectResult
 
+    /**
+     *
+     *
+     * Returns a list of summary information about the objects in the specified
+     * buckets. List results are *always* returned in lexicographic
+     * (alphabetical) order.
+     *
+     *
+     *
+     * Because buckets can contain a virtually unlimited number of keys, the
+     * complete results of a list query can be extremely large. To manage large
+     * result sets, Amazon S3 uses pagination to split them into multiple
+     * responses.
+     *
+     *
+     * The total number of keys in a bucket doesn't substantially affect list
+     * performance.
+     *
+     *
+     * @param bucketName The name of the Amazon S3 bucket to list.
+     * @return A listing of the objects in the specified bucket, along with any
+     * other associated information, such as common prefixes (if a
+     * delimiter was specified), the original request parameters, etc.
+     * @throws AwsException If any errors are encountered in the client
+     * while making the request or handling the response.
+     * @see AWSS3.listObjects
+     */
+    suspend fun listObjects(bucketName: String): ListObjectsResult
+
     class Builder() {
         fun accessKey(accessKey: String): Builder
         fun secretKey(secretKey: String): Builder
         fun setEndpoint(endpoint: String): Builder
-        fun build(): AwsS3
+        fun build(): AWSS3
     }
 }
