+++ aws-s3/src/commonMain/kotlin/com/estivensh4/aws_s3/CopyObjectResult.kt
@@ -6,9 +6,9 @@ package com.estivensh4.aws_s3
 
 import kotlinx.datetime.Instant
 
-data class PutObjectResult(
+data class CopyObjectResult(
     val versionId: String?,
     val eTag: String?,
     val expirationTime: Instant?,
-    val contentMd5: String?,
+    val sseCustomerKeyMd5: String?,
 )
