+++ aws-s3/src/iosMain/kotlin/com/estivensh4/aws_s3/util/AWSS3Ext.kt
@@ -1,3 +1,7 @@
+/*
+ * Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+ */
+
 package com.estivensh4.aws_s3.util
 
 import cocoapods.AWSS3.AWSS3
