+++ build-logic/src/main/kotlin/publication-convention.gradle.kts
@@ -1,3 +1,7 @@
+/*
+ * Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+ */
+
 plugins {
     id("org.gradle.maven-publish")
 }
