+++ doc/Writerside/cfg/platforms.xml
@@ -1,4 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+  -->
+
 <platforms>
 <platform id="primary">
     <src>keymap.xml</src>
