+++ example/androidapp/src/main/res/drawable/ic_launcher_background.xml
@@ -1,4 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+  -->
+
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
     android:width="108dp"
     android:height="108dp"
