+++ example/androidapp/src/main/res/drawable/ic_launcher_foreground.xml
@@ -1,3 +1,7 @@
+<!--
+  ~ Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+  -->
+
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:aapt="http://schemas.android.com/aapt"
     android:width="108dp"
