+++ example/androidapp/src/main/res/values/strings.xml
@@ -1,3 +1,7 @@
+<!--
+  ~ Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+  -->
+
 <resources>
     <string name="app_name">androidApp</string>
 </resources>
