+++ example/androidapp/src/main/res/values/themes.xml
@@ -1,4 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+  -->
+
 <resources>
 
     <style name="Theme.Awskmp" parent="android:Theme.Material.Light.NoActionBar" />
