+++ example/gradle/wrapper/gradle-wrapper.properties
@@ -1,3 +1,7 @@
+#
+# Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+#
+
 #Thu Nov 09 01:52:10 COT 2023
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
