+++ example/wearapp/src/main/res/values-round/strings.xml
@@ -1,3 +1,7 @@
+<!--
+  ~ Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+  -->
+
 <resources>
     <string name="hello_world">From the Round world,\nHello, %1$s!</string>
 </resources>
