+++ example/webApp/src/jsMain/resources/index.html
@@ -1,5 +1,5 @@
 <!--
-  ~ Copyright 2022 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+  ~ Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
   -->
 
 <!DOCTYPE html>
