+++ gradle.properties
@@ -1,3 +1,7 @@
+#
+# Copyright 2023 estiven. Use of this source code is governed by the Apache 2.0 license.
+#
+
 #Gradle
 org.gradle.jvmargs=-Xmx2048M -Dfile.encoding=UTF-8 -Dkotlin.daemon.jvm.options\="-Xmx2048M"
