+++ gradle/aws.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-aws = "0.4.4"
+aws = "0.5.0"
 
 [libraries]
 aws-common = { module = "io.github.estivensh4:aws-common", version.ref = "aws" }
